refactor: update DonateModal to DonateScreen
