+++ src/config/navigation/index.tsx
@@ -12,7 +12,7 @@ import NetworkProvider from "contexts/networkContext";
 import CryptoPaymentProvider from "contexts/cryptoPaymentContext";
 import CausesProvider from "contexts/causesContext";
 import CurrentUserProvider from "contexts/currentUserContext";
-import DonateModal from "screens/donations/DonateModal";
+import DonateScreen from "screens/donations/DonateScreen";
 import NotFoundScreen from "screens/NotFoundScreen";
 import CausesScreen from "screens/donations/CausesScreen";
 import ProfileScreen from "screens/users/ProfileScreen";
@@ -81,14 +81,23 @@ function RootNavigator() {
         name="PaymentScreen"
         component={PaymentScreen}
         options={{
-            headerTintColor: theme.colors.orange40,
-            headerTitle: "",
-            headerBackTitleVisible: false,
+          headerTintColor: theme.colors.orange40,
+          headerTitle: "",
+          headerBackTitleVisible: false,
+        }}
+      />
+
+      <Stack.Screen
+        name="DonateScreen"
+        component={DonateScreen}
+        options={{
+          headerTintColor: theme.colors.orange40,
+          headerTitle: "",
+          headerBackTitleVisible: false,
         }}
       />
 
       <Stack.Group screenOptions={{ presentation: "modal" }}>
-        <Stack.Screen name="DonateModal" component={DonateModal} />
         <Stack.Screen
           name="CommunityAddModal"
           component={CommunityAddScreen}
